Move PreLaunchChecklist and related code out from the step 3 of MC setup
