+++ js/src/setup-mc/setup-stepper/setup-free-listings/index.js
@@ -75,11 +75,6 @@ const SetupFreeListings = ( props ) => {
 					free_shipping_threshold: settings.free_shipping_threshold,
 					shipping_time: settings.shipping_time,
 					tax_rate: settings.tax_rate,
-					website_live: settings.website_live,
-					checkout_process_secure: settings.checkout_process_secure,
-					payment_methods_visible: settings.payment_methods_visible,
-					refund_tos_visible: settings.refund_tos_visible,
-					contact_info_visible: settings.contact_info_visible,
 				} }
 				validate={ handleValidate }
 				onSubmitCallback={ handleSubmitCallback }
