+++ js/src/setup-mc/setup-stepper/store-requirements/pre-launch-checklist/index.scss
